feat(xp): redesign SP economy — role-based skill points replace 1:1 XP
Replace the trivial 1:1 SP-from-XP system with a scarce, role-based
economy where SP is earned through quality outcomes and penalized for
failures. Target: ~2-5 SP/week for consistent quality work.

**SP Rules by Role:**
- Developer: PR merged +0.5, review given +0.25, BUD shipped +1.0,
  quality>80 +0.5, testing bug -0.25, prod bug -1.0
- QA: every 5 testing bugs +1.0, prod bug found +0.5, false positive -0.25
- PM: BUD shipped +0.5, prompt approval +0.25, BUD discarded -0.5
- Tech Lead: review done +0.25, arch approved +0.25, missed prod bug -0.5
- All: 14-day streak +1.0, 30-day streak +2.0

**Foundation:**
- Migration: skill_points Integer → Numeric(10,2) with reset to 0
- New sp_rules.py (constants) and sp_service.py (award/penalize/role)
- Dedup via source_ref, floor at 0, WebSocket publish

**Triggers wired into existing flows:**
- github_webhook_handler: PR merged, code review submitted
- bud_closure: BUD shipped to PROD
- xp_service: quality bonus, streak milestones
- bugs.py: developer penalty + QA reward on bug creation

**Frontend:**
- Fractional SP display in XPProfileCard (visible on Profile + CharacterSelect)
- Comprehensive role-specific SP guide in XPInfoPanel
- Knight character lock raised to level 3

+++ backend/app/schemas/xp.py
@@ -17,7 +17,7 @@ class XPProfileRead(BaseModel):
     streak_best: int
     unlocked_characters: list[str]
     unlocked_accessories: list[str]
-    skill_points: int = 0
+    skill_points: float = 0.0
     house_level: int = 2
     vehicle_unlocks: list[str] = []
 
@@ -48,7 +48,7 @@ class VehicleUnlockResponse(BaseModel):
     """Response after unlocking a vehicle."""
 
     success: bool
-    remaining_skill_points: int
+    remaining_skill_points: float
     vehicle_unlocks: list[str]
 
 
@@ -62,7 +62,7 @@ class HouseUpgradeResponse(BaseModel):
     """Response after upgrading house."""
 
     success: bool
-    remaining_skill_points: int
+    remaining_skill_points: float
     house_level: int
